requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
google-auth==2.23.4
google-generativeai==0.3.2
gunicorn==21.2.0
